requirements.txt
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
slowapi>=0.1.9
prometheus-client>=0.17.0
psutil>=5.9.0
datasets>=2.13.0
evaluate>=0.4.0
scikit-learn>=1.3.0
sentencepiece>=0.1.99
safetensors>=0.3.0
python-dotenv>=1.0.0
PyJWT>=2.7.0
jinja2>=3.1.2
websockets>=11.0
